+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * **Breaking Change** play scales for two octaves rather than one.
 * Support scales "concieved of ... using a division of the perfect fourth interval"
+* Add `-l` option to read from legends.xml, and `-L` to list forms defined there.
 
 # 0.0.0